requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-dotenv
openai
firebase-admin
